Add anonymous tracking
Extended to do two things; prompt for opt in/out of
collection, and sending the POST. After installing/upgrading the plugin, the
user is prompted to opt in/out of the collection with a simple y|n option
(default y). Both options set a config file in the existing ~/.heroku/ folder,
with a setting opt-in:true|false value.

For every command ran in the plugin, a background thread will check the this
config file and the value in it. If yes, the command is posted to the Sinatra
app. A thread is used to help ensure this process doesn't interfere with normal
usage. The plugin does not wait for or ensure the POST completes.