Add mean average query execution time to outliers
This commit adds a mean average of query execution time to
the outliers query, hidden behind a flag.

The rationale is that mean average execution time is to provide
a quicker way to determine queries that are slow, rather than
having to determine the ratio manually. It also provides
a needle to move when improving queries that appear on the
outliers report.

Hiding this behind a flag means existing output parsers won't
break.

* Add optional average execution time to outliers report
* Add flag to opt into average execution time in report