Expand pg:locks query
This commit expands the pg:locks query to better handle expanded
lock types, and provide richer information about the locks.

The main downside is that by including more lock types, there is
a lot more information. For example, obtaining an `AccessExclusiveLock`
on a table (as part of `DROP TABLE`), grabs locks on sequences, TOAST
data and indexes as well as the table itself, as well as locks on
`pg_type` and other system tables, that aren't easily understood.

This is best highlighted with the updated example in the README.

In order to help account for this, new columns have been added to
help identify the lock types as well as lock modes, and provide
some further information. This might be too much, so feedback
is deeply appreciated.

* Expand locks query to include `AccessExclusiveLock` and
`RowExclusiveLock`
* Add further metadata about acquired locks to output
* Update example

Fixes #149